generate the entire HTML file on the server side
