+++ src/meshcat/servers/zmqserver.py
@@ -71,9 +71,13 @@ class WebSocketHandler(tornado.websocket.WebSocketHandler):
         print("closed:", self, file=sys.stderr)
 
 
-def create_base64(data):
-    """Create a base64 encoded version of the Three.js data that can be embedded."""
-    return base64.b64encode(data)
+def create_command(data) -> str:
+    """Encode the drawing command into a Javascript fetch() command for display."""
+    return """
+fetch("data:application/octet-binary;base64,{}")
+    .then(res => res.arrayBuffer())
+    .then(buffer => viewer.handle_command_bytearray(new Uint8Array(buffer)));
+    """.format(base64.b64encode(data).decode("utf-8"))
 
 
 class ZMQWebSocketBridge(object):
@@ -142,13 +146,48 @@ class ZMQWebSocketBridge(object):
         elif cmd == "get_scene":
             # when the server gets this command, return the tree
             # as a series of msgpack-backed binary blobs
-            output = b""
+            drawing_commands = ""
             for node in walk(self.tree):
                 if node.object is not None:
-                    output += create_base64(node.object)
+                    drawing_commands += create_command(node.object)
                 if node.transform is not None:
-                    output += create_base64(node.transform)
-            self.zmq_socket.send(output)
+                    drawing_commands += create_command(node.transform)
+
+            # now that we have the drawing commands, generate the full
+            # HTML that we want to generate, including the javascript assets
+            mainminjs_path = os.path.join(VIEWER_ROOT, "main.min.js")
+            mainminjs_src = ""
+            with open(mainminjs_path, "r") as f:
+                mainminjs_src = f.readlines()
+            mainminjs_src = "".join(mainminjs_src)
+
+            html = """
+                <!DOCTYPE html>
+                <html>
+                    <head> <meta charset=utf-8> <title>MeshCat</title> </head>
+                    <body>
+                        <div id="meshcat-pane">
+                        </div>
+                        <script>
+                            {mainminjs}
+                        </script>
+                        <script>
+                            var viewer = new MeshCat.Viewer(document.getElementById("meshcat-pane"));
+                            {commands}
+                        </script>
+                         <style>
+                            body {{margin: 0; }}
+                            #meshcat-pane {{
+                                width: 100vw;
+                                height: 100vh;
+                                overflow: hidden;
+                            }}
+                        </style>
+                        <script id="embedded-json"></script>
+                    </body>
+                </html>
+            """.format(mainminjs=mainminjs_src, commands=drawing_commands)
+            self.zmq_socket.send(html.encode('utf-8'))
         else:
             self.zmq_socket.send(b"error: unrecognized comand")
